refactor: PR 2 — common 정리 (web/persistence/image 재배치)
- common/web 신설: response/filter + GlobalExceptionHandler 통합
- common/web/swagger 중첩: 응답 어노테이션 4종 + OptionalAuth 분리
- common/persistence/BaseEntity 이동
- common/image 신설: entity (Image, ImageTargetType), persistence (ImageRepository)
- ImageQueryService → ImageUrlReader 리네임 + @Transactional 제거
  (Spring Data repo 호출 자체가 암묵 transactional, B3/C2 룰 통과)
- ArchUnitConstants.FEATURES 에서 image 제거 (common 잔류)
- CommonPackageRuleTest.X2 활성 (common.web/config 도메인 의존 금지)
