+++ src/main/kotlin/com/team2/server/auth/controller/DevTokenController.kt
@@ -3,7 +3,7 @@ package com.team2.server.auth.controller
 import com.team2.server.auth.jwt.JwtTokenProvider
 import com.team2.server.common.exception.BusinessException
 import com.team2.server.common.exception.ErrorCode
-import com.team2.server.common.response.ApiResponse
+import com.team2.server.common.web.ApiResponse
 import com.team2.server.user.entity.AuthProvider
 import com.team2.server.user.entity.User
 import com.team2.server.user.repository.UserRepository
