+++ src/main/kotlin/com/team2/server/chat/controller/ChatApi.kt
@@ -6,10 +6,10 @@ import com.team2.server.auth.principal.UserPrincipal
 import com.team2.server.chat.dto.ChatMessageResponse
 import com.team2.server.chat.dto.EnterRealtimePartyRequest
 import com.team2.server.chat.dto.SendChatMessageRequest
-import com.team2.server.common.response.ApiResponse
-import com.team2.server.common.response.ErrorResponse
-import com.team2.server.common.swagger.AuthErrorResponses
-import com.team2.server.common.swagger.InternalServerErrorResponse
+import com.team2.server.common.web.ApiResponse
+import com.team2.server.common.web.ErrorResponse
+import com.team2.server.common.web.swagger.AuthErrorResponses
+import com.team2.server.common.web.swagger.InternalServerErrorResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.enums.ParameterIn
