+++ src/main/kotlin/com/team2/server/common/image/entity/ImageTargetType.kt
@@ -1,4 +1,4 @@
-package com.team2.server.common.entity
+package com.team2.server.common.image.entity
 
 enum class ImageTargetType {
     ROLLING_PAPER_WRAPPER,
