+++ src/main/kotlin/com/team2/server/common/web/swagger/ValidationErrorResponse.kt
@@ -1,6 +1,6 @@
-package com.team2.server.common.swagger
+package com.team2.server.common.web.swagger
 
-import com.team2.server.common.response.ErrorResponse
+import com.team2.server.common.web.ErrorResponse
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.ExampleObject
 import io.swagger.v3.oas.annotations.media.Schema
@@ -9,8 +9,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 @Target(AnnotationTarget.FUNCTION)
 @Retention(AnnotationRetention.RUNTIME)
 @ApiResponse(
-    responseCode = "500",
-    description = "서버 내부 오류",
+    responseCode = "400",
+    description = "입력값 검증 실패",
     content = [
         Content(
             mediaType = "application/json",
@@ -19,10 +19,10 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
                 ExampleObject(
                     value = """
                         {
-                          "status": 500,
+                          "status": 400,
                           "error": {
-                            "code": "INTERNAL_SERVER_ERROR",
-                            "message": "서버 내부 오류가 발생했습니다"
+                            "code": "VALIDATION_ERROR",
+                            "message": "nickname: 닉네임은 필수입니다"
                           }
                         }
                     """,
@@ -31,4 +31,4 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
         ),
     ],
 )
-annotation class InternalServerErrorResponse
+annotation class ValidationErrorResponse
