+++ src/main/kotlin/com/team2/server/party/controller/CharacterController.kt
@@ -1,6 +1,6 @@
 package com.team2.server.party.controller
 
-import com.team2.server.common.response.ApiResponse
+import com.team2.server.common.web.ApiResponse
 import com.team2.server.party.dto.CharacterResponse
 import com.team2.server.party.usecase.GetCharactersUseCase
 import org.springframework.web.bind.annotation.GetMapping
