+++ src/main/kotlin/com/team2/server/party/controller/MePartyApi.kt
@@ -1,9 +1,9 @@
 package com.team2.server.party.controller
 
 import com.team2.server.auth.principal.UserPrincipal
-import com.team2.server.common.response.ApiResponse
-import com.team2.server.common.swagger.AuthErrorResponses
-import com.team2.server.common.swagger.InternalServerErrorResponse
+import com.team2.server.common.web.ApiResponse
+import com.team2.server.common.web.swagger.AuthErrorResponses
+import com.team2.server.common.web.swagger.InternalServerErrorResponse
 import com.team2.server.party.dto.UpcomingPartyResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
