+++ src/main/kotlin/com/team2/server/party/entity/RealtimeParticipantProfile.kt
@@ -1,6 +1,6 @@
 package com.team2.server.party.entity
 
-import com.team2.server.common.entity.BaseEntity
+import com.team2.server.common.persistence.BaseEntity
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
