+++ src/main/kotlin/com/team2/server/rollingpaper/controller/RollingPaperController.kt
@@ -1,7 +1,7 @@
 package com.team2.server.rollingpaper.controller
 
 import com.team2.server.auth.principal.UserPrincipal
-import com.team2.server.common.response.ApiResponse
+import com.team2.server.common.web.ApiResponse
 import com.team2.server.rollingpaper.dto.CreateRollingPaperRequest
 import com.team2.server.rollingpaper.dto.CreateRollingPaperResponse
 import com.team2.server.rollingpaper.dto.ParticipantRollingPaperListResponse
