+++ src/main/kotlin/com/team2/server/rollingpaper/controller/RollingPaperOwnerApi.kt
@@ -1,10 +1,10 @@
 package com.team2.server.rollingpaper.controller
 
 import com.team2.server.auth.principal.UserPrincipal
-import com.team2.server.common.response.ApiResponse
-import com.team2.server.common.response.ErrorResponse
-import com.team2.server.common.swagger.AuthErrorResponses
-import com.team2.server.common.swagger.InternalServerErrorResponse
+import com.team2.server.common.web.ApiResponse
+import com.team2.server.common.web.ErrorResponse
+import com.team2.server.common.web.swagger.AuthErrorResponses
+import com.team2.server.common.web.swagger.InternalServerErrorResponse
 import com.team2.server.rollingpaper.dto.OwnerRollingPaperDetailResponse
 import com.team2.server.rollingpaper.dto.OwnerRollingPaperListResponse
 import io.swagger.v3.oas.annotations.Operation
