+++ src/main/kotlin/com/team2/server/rollingpaper/controller/RollingPaperOwnerController.kt
@@ -1,7 +1,7 @@
 package com.team2.server.rollingpaper.controller
 
 import com.team2.server.auth.principal.UserPrincipal
-import com.team2.server.common.response.ApiResponse
+import com.team2.server.common.web.ApiResponse
 import com.team2.server.rollingpaper.dto.OwnerRollingPaperDetailResponse
 import com.team2.server.rollingpaper.dto.OwnerRollingPaperListResponse
 import com.team2.server.rollingpaper.usecase.GetRollingPaperDetailUseCase
