+++ src/main/kotlin/com/team2/server/rollingpaper/usecase/GetRollingPaperListUseCase.kt
@@ -1,9 +1,9 @@
 package com.team2.server.rollingpaper.usecase
 
-import com.team2.server.common.entity.ImageTargetType
 import com.team2.server.common.exception.BusinessException
 import com.team2.server.common.exception.ErrorCode
-import com.team2.server.common.service.ImageQueryService
+import com.team2.server.common.image.entity.ImageTargetType
+import com.team2.server.common.image.persistence.ImageUrlReader
 import com.team2.server.party.entity.Party
 import com.team2.server.party.entity.PartyOption
 import com.team2.server.party.repository.PartyInviteRepository
@@ -25,7 +25,7 @@ class GetRollingPaperListUseCase(
     private val partyInviteRepository: PartyInviteRepository,
     private val partyRepository: PartyRepository,
     private val rollingPaperRepository: RollingPaperRepository,
-    private val imageQueryService: ImageQueryService,
+    private val imageUrlReader: ImageUrlReader,
 ) {
     @Transactional(readOnly = true)
     fun getParticipantList(
@@ -132,7 +132,7 @@ class GetRollingPaperListUseCase(
     ): Long = ((page - 1) * PAGE_SIZE + index + 1).toLong()
 
     private fun findImageUrlByWrapperId(rollingPapers: List<RollingPaper>): Map<Long, String> =
-        imageQueryService.findFirstImageUrlByTargetIds(
+        imageUrlReader.findFirstImageUrlByTargetIds(
             ImageTargetType.ROLLING_PAPER_WRAPPER,
             rollingPapers.map { it.wrapper.id },
         )
