+++ src/test/kotlin/com/team2/server/architecture/ArchUnitConstants.kt
@@ -3,7 +3,7 @@ package com.team2.server.architecture
 object ArchUnitConstants {
     const val BASE_PACKAGE = "com.team2.server"
 
-    val FEATURES = listOf("auth", "user", "party", "image", "chat", "rollingpaper")
+    val FEATURES = listOf("auth", "user", "party", "chat", "rollingpaper")
 
     const val API = "..api.."
     const val USECASE = "..application.usecase.."
