+++ src/test/kotlin/com/team2/server/architecture/CommonPackageRuleTest.kt
@@ -2,7 +2,6 @@ package com.team2.server.architecture
 
 import com.tngtech.archunit.core.importer.ImportOption
 import com.tngtech.archunit.junit.AnalyzeClasses
-import com.tngtech.archunit.junit.ArchIgnore
 import com.tngtech.archunit.junit.ArchTest
 import com.tngtech.archunit.lang.ArchRule
 import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses
@@ -12,7 +11,7 @@ import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses
  *
  * 활성화 Phase:
  *  - X1: Phase 0 (즉시)
- *  - X2: Phase 1 (common 정리 후)
+ *  - X2: PR 2 (common 정리 후)
  */
 @AnalyzeClasses(
     packages = ["com.team2.server"],
@@ -32,9 +31,8 @@ class CommonPackageRuleTest {
             .resideInAnyPackage(*featurePackages)
             .`as`("common 패키지는 feature 패키지 의존 금지")
 
-    /** X2: domain 은 common.web/config/filter/response 의존 금지 (exception/persistence/entity 만 허용). */
+    /** X2: domain 은 common.web/config 의존 금지 (exception/persistence/image 만 허용). */
     @ArchTest
-    @ArchIgnore(reason = "Phase 1: common 정리 후 활성화")
     val domainShouldNotDependOnWebLayerCommon: ArchRule =
         noClasses()
             .that()
@@ -44,7 +42,5 @@ class CommonPackageRuleTest {
             .resideInAnyPackage(
                 "..common.web..",
                 "..common.config..",
-                "..common.filter..",
-                "..common.response..",
-            ).`as`("Domain 은 common 의 web/config/filter/response 의존 금지")
+            ).`as`("Domain 은 common 의 web/config 의존 금지")
 }
